[BE] [#214] Service 인터페이스 제거 및 Service 클래스 구조로 통합 (#215)
* user service interface 삭제, service impl를 단독 class로 수정

* refactor: playlist, comment, gpt service interface 삭제후 service impl class 단일 클래스로 수정
